refactoring, link up button in about window
